changed type from long to int

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -9,17 +9,17 @@
 int main(void)
 {
 
-	unsigned long int i;
+	int i;
 
-	unsigned long int n;
+	int n;
 
-	unsigned long int t1 = 1;
+	int t1 = 1;
 
-	unsigned long int t2 = 2;
+	int t2 = 2;
 
-	unsigned long int nextTerm;
+	int nextTerm;
 
-	unsigned long int even = 0;
+	int even = 0;
 
 	while (t1 < 4000000 && t2 < 4000000)
 	{
